Moved command and query models to Models project as that is more appropriate. This makes RCommon.Models a dependency of RCommon.ApplicationServices but this is probably the way in needs to be moving forward.
