+++ Examples/ApplicationServices/Examples.ApplicationServices.CQRS/Program.cs
@@ -5,7 +5,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using RCommon;
 using RCommon.ApplicationServices;
-using RCommon.ApplicationServices.ExecutionResults;
 using RCommon.Caching;
 using RCommon.FluentValidation;
 using System.Diagnostics;
