+++ Examples/ApplicationServices/Examples.ApplicationServices.CQRS/TestApplicationService.cs
@@ -1,6 +1,6 @@
 ﻿using RCommon.ApplicationServices.Commands;
-using RCommon.ApplicationServices.ExecutionResults;
 using RCommon.ApplicationServices.Queries;
+using RCommon.Models.ExecutionResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
