+++ Examples/ApplicationServices/Examples.ApplicationServices.CQRS/TestCommand.cs
@@ -1,5 +1,5 @@
-﻿using RCommon.ApplicationServices.Commands;
-using RCommon.ApplicationServices.ExecutionResults;
+﻿using RCommon.Models.Commands;
+using RCommon.Models.ExecutionResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
