+++ Examples/ApplicationServices/Examples.ApplicationServices.CQRS/TestCommandHandler.cs
@@ -1,6 +1,6 @@
 ﻿using RCommon;
 using RCommon.ApplicationServices.Commands;
-using RCommon.ApplicationServices.ExecutionResults;
+using RCommon.Models.ExecutionResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
