+++ Examples/ApplicationServices/Examples.ApplicationServices.CQRS/TestQuery.cs
@@ -1,4 +1,4 @@
-﻿using RCommon.ApplicationServices.Queries;
+﻿using RCommon.Models.Queries;
 using System;
 using System.Collections.Generic;
 using System.Linq;
