+++ Examples/Validation/Examples.Validation.FluentValidation/Program.cs
@@ -4,7 +4,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using RCommon;
 using RCommon.ApplicationServices;
-using RCommon.ApplicationServices.ExecutionResults;
 using RCommon.FluentValidation;
 using System.Diagnostics;
