+++ Examples/Validation/Examples.Validation.FluentValidation/TestApplicationService.cs
@@ -1,5 +1,4 @@
 ﻿using RCommon.ApplicationServices.Commands;
-using RCommon.ApplicationServices.ExecutionResults;
 using RCommon.ApplicationServices.Queries;
 using RCommon.ApplicationServices.Validation;
 using System;
