+++ Src/RCommon.ApplicationServices/Commands/ICommandBus.cs
@@ -21,7 +21,8 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using RCommon.ApplicationServices.ExecutionResults;
+using RCommon.Models.Commands;
+using RCommon.Models.ExecutionResults;
 using System.Threading;
 using System.Threading.Tasks;
