+++ Src/RCommon.ApplicationServices/Commands/ICommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using RCommon.ApplicationServices.ExecutionResults;
+﻿using RCommon.Models.Commands;
+using RCommon.Models.ExecutionResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
