+++ Src/RCommon.ApplicationServices/CqrsBuilderExtensions.cs
@@ -24,8 +24,10 @@
 using Microsoft.Extensions.DependencyInjection;
 using RCommon.ApplicationServices;
 using RCommon.ApplicationServices.Commands;
-using RCommon.ApplicationServices.ExecutionResults;
 using RCommon.ApplicationServices.Queries;
+using RCommon.Models.Commands;
+using RCommon.Models.ExecutionResults;
+using RCommon.Models.Queries;
 using System;
 using System.Collections.Generic;
 using System.Linq;
