+++ Src/RCommon.ApplicationServices/Queries/IQueryHandler.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using RCommon.Models.Queries;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
