+++ Src/RCommon.ApplicationServices/RCommon.ApplicationServices.csproj
@@ -7,6 +7,7 @@
   <ItemGroup>
     <ProjectReference Include="..\RCommon.Caching\RCommon.Caching.csproj" />
     <ProjectReference Include="..\RCommon.Core\RCommon.Core.csproj" />
+    <ProjectReference Include="..\RCommon.Models\RCommon.Models.csproj" />
   </ItemGroup>
 	<ItemGroup>
 	  <Folder Include="Caching\" />
