+++ Src/RCommon.Models/Commands/ICommand.cs
@@ -1,11 +1,11 @@
-﻿using RCommon.ApplicationServices.ExecutionResults;
+﻿using RCommon.Models.ExecutionResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RCommon.ApplicationServices.Commands
+namespace RCommon.Models.Commands
 {
     public interface ICommand
     {
