+++ Src/RCommon.Models/ExecutionResults/IExecutionResult.cs
@@ -21,28 +21,10 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System.Collections.Generic;
-using System.Linq;
-
-namespace RCommon.ApplicationServices.ExecutionResults
+namespace RCommon.Models.ExecutionResults
 {
-    public class FailedExecutionResult : ExecutionResult
+    public interface IExecutionResult
     {
-        public IReadOnlyCollection<string> Errors { get; }
-
-        public FailedExecutionResult(
-            IEnumerable<string> errors)
-        {
-            Errors = (errors ?? Enumerable.Empty<string>()).ToList();
-        }
-            
-        public override bool IsSuccess { get; } = false;
-
-        public override string ToString()
-        {
-            return Errors.Any()
-                ? $"Failed execution due to: {string.Join(", ", Errors)}"
-                : "Failed execution";
-        }
+        bool IsSuccess { get; }
     }
 }
