+++ Src/RCommon.Models/ExecutionResults/SuccessExecutionResult.cs
@@ -21,10 +21,15 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-namespace RCommon.ApplicationServices.ExecutionResults
+namespace RCommon.Models.ExecutionResults
 {
-    public interface IExecutionResult
+    public class SuccessExecutionResult : ExecutionResult
     {
-        bool IsSuccess { get; }        
+        public override bool IsSuccess { get; } = true;
+
+        public override string ToString()
+        {
+            return "Successful execution";
+        }
     }
 }
