+++ Tests/RCommon.TestBase.Data/TestRepository.cs
@@ -26,7 +26,7 @@ namespace RCommon.TestBase.Data
             _entityDeleteActions = new List<Action<DbContext>>();
         }
 
-        public TestRepository(ServiceProvider serviceProvider)
+        public TestRepository(IServiceProvider serviceProvider)
         {
             _dataStoreFactory = serviceProvider.GetService<IDataStoreFactory>();
             _context = _dataStoreFactory.Resolve<RCommonDbContext>("TestDbContext");
